stream-reactor commit

+++ download_kafka.sh
@@ -8,23 +8,27 @@ DOWNLOAD_DIR=~/Documents
 SH_DIR=${DOWNLOAD_DIR}/shell_script
 
 ## Download Kafka.tgz file
-wget -d -P ${DOWNLOAD_DIR} ${KAFKA_URL}/${KAFKA_VER}/${DOWNLOAD_FILE}
-wget -d -P ${DOWNLOAD_DIR} ${KAFKA_URL}/${KAFKA_VER}/${DOWNLOAD_FILE}.asc
-wget -d -P ${DOWNLOAD_DIR} ${KAFKA_URL}/${KAFKA_VER}/${DOWNLOAD_FILE}.sha512
-echo -e ">> Download Complete.\n"
+if [ ! -e ${KAFKA_URL}/${KAFKA_VER}/${DOWNLOAD_FILE} ]; then
+    wget -d -P ${DOWNLOAD_DIR} ${KAFKA_URL}/${KAFKA_VER}/${DOWNLOAD_FILE}
+    wget -d -P ${DOWNLOAD_DIR} ${KAFKA_URL}/${KAFKA_VER}/${DOWNLOAD_FILE}.asc
+    wget -d -P ${DOWNLOAD_DIR} ${KAFKA_URL}/${KAFKA_VER}/${DOWNLOAD_FILE}.sha512
+    echo -e ">> Download Complete.\n"
 
-## Verify the integrity
-cd ${DOWNLOAD_DIR}
-# gpg --import ${DOWNLOAD_FILE}.asc
-gpg --verify ${DOWNLOAD_FILE}.asc ${DOWNLOAD_FILE}
+    ## Verify the integrity
+    cd ${DOWNLOAD_DIR}
+    # gpg --import ${DOWNLOAD_FILE}.asc
+    gpg --verify ${DOWNLOAD_FILE}.asc ${DOWNLOAD_FILE}
 
-echo -e "\n[ Calculated Hash Value ]"
-gpg --print-md SHA512 ${DOWNLOAD_FILE}
+    echo -e "\n[ Calculated Hash Value ]"
+    gpg --print-md SHA512 ${DOWNLOAD_FILE}
 
-echo "[ Known Hash Value ]"
-cat ${DOWNLOAD_FILE}.sha512
+    echo "[ Known Hash Value ]"
+    cat ${DOWNLOAD_FILE}.sha512
 
-rm ${DOWNLOAD_FILE}.*
+    rm ${DOWNLOAD_FILE}.*
+else
+    echo -e ">> kafka.tgz file exists already."
+fi
 
 ## Unzip kafka.tgz
 tar xzf ${DOWNLOAD_FILE} ${DOWNLOAD_DIR}
